exponiendo la pai

+++ src/main/java/com/examplemarket/persistence/entity/ComprasProductoPK.java
@@ -6,26 +6,25 @@ import java.io.Serializable;
 
 @Embeddable
 public class ComprasProductoPK implements Serializable {
-
     @Column(name = "id_compra")
-    private Long idCompra;
+    private Integer idCompra;
 
     @Column(name = "id_producto")
-    private Long idProducto;
+    private Integer idProducto;
 
-    public Long getIdCompra() {
+    public Integer getIdCompra() {
         return idCompra;
     }
 
-    public void setIdCompra(Long idCompra) {
+    public void setIdCompra(Integer idCompra) {
         this.idCompra = idCompra;
     }
 
-    public Long getIdProducto() {
+    public Integer getIdProducto() {
         return idProducto;
     }
 
-    public void setIdProducto(Long idProducto) {
+    public void setIdProducto(Integer idProducto) {
         this.idProducto = idProducto;
     }
 }
